Docker build image action fixes
* Print complete traceback on exceptions for better visibility.
* JSON output from docker daemon is inconsistent depending on whether
a base image is already existing or is being downloaded. You can either
get {'stream': 'some message'} or {'status': 'STATUS', other fields...}.
So this action now prints the entire JSON output.
* While using chunked transfer (stream=True), sometimes we get a ReadTimeoutError.
In most cases, it is temporary because docker daemon is loaded. So
we just ignore the ReadTimeoutError and retry. Note this is not ideal
but as a client there is little we could do.