Allow multiple exclamation marks while honking Tesla
Since honking is a very emotional process, and people often have to
channel their intense feelings into it, it makes very little sense to
have a `?` quantifier in the honking alias: that would only allow for a
mere `honk-honk`, or a brief exclamative `honk-honk!`.

Changing the quantifier to `*` gives people the chance to express all
the feelings they want to express. `honk-honk!!!!!!!!!!!!!!!!!!!!!!!`